feat(buftype): add possibility of saving files locally
Necessary for building pdfs locally. The feature is added as an opt-in
setting. The added variable 'g:AirLatexBuftype' has default value 'NOFILE'
which has the usual behaviour (sets 'buftype' to 'nofile'). The value
'NORMAL' creates a normal buffer (sets 'buftype' to '') which allows for
local saving of the files. Any other value will not be accepted and will
have the same effect as the default.